Fixing indentation error in subscribe that allowed an attempt to sort dependencies when no receivers are specified.
Adding unittests for topological sort that ensure the result is
properly ordered and that a ValueError is raised for a cycle only if
it causes there to be no root nodes in the dependency graph.

NB: The tsort function blissfully ignores over all other cycles.  On
the one hand the results may be unpredictable so it might be best to
fail on any cycle, while on the other hand too much strictness may be
unpleasant.

Adding unittests for the Pangler class that ensure the subscribe
method's 'receivers' keyword populates a 'dependencies' dict as well
as an 'order' list, and that a series of chained events operate in the
correct sequence and on the correct parameters.